requirements.txt
bcrypt==4.3.0
cffi==1.17.1
cryptography==45.0.4
decorator==5.2.1
deprecated==1.2.18
fabric==3.2.2
invoke==2.2.0
jsonpath-ng==1.7.0
paramiko==3.5.1
ply==3.11
pycparser==2.22
pynacl==1.5.0
python-slugify==8.0.4
text-unidecode==1.3
wrapt==1.17.2
